The docount function should have been status.  All size variables (iSize) should have been "size_t" and not "int".

+++ rbtree.h
@@ -12,18 +12,18 @@ typedef struct {
     int (*cmp)(void *ctx, const void *a, const void *b);
     RB_NODE *(*newnode)(void *ctx, const void *data);
     void (*delnode)(void *ctx, RB_NODE *node);
-    int iSize;
+    size_t iSize;
     } RB_TYPE;
 
 void rb_type_create(
-    RB_TYPE *rbt,int iSize, void *ctx,
+    RB_TYPE *rbt,size_t iSize, void *ctx,
     int (*cmp)(void*,const void *,const void *),/* Required */
     RB_NODE *(*newnode)(void*,const void*),     /* Defaults to malloc */
     void (*delnode)(void*,RB_NODE*) );          /* Defaults to free */
 
 void rb_free( const RB_TYPE *rbt, RB_TREE *root );
 
-int rb_size( RB_TREE *root );
+size_t rb_size( RB_TREE *root );
 
 RB_NODE *rb_search(const RB_TYPE *rbt, RB_NODE *n, void *data);
 int rb_insert( const RB_TYPE *rbt, RB_TREE *root, void *data );
